Moved and renamed 'apiHandleError' and updated its imports
